requirements.txt
# Django & Channels
Django>=5.0
daphne>=4.0
channels>=4.0

# REST API
djangorestframework>=3.14
drf-spectacular>=0.27

# ML and Data Processing
pandas>=2.0
numpy>=1.24
scikit-learn>=1.4
matplotlib

# Database
psycopg2-binary

# PDF Generation
weasyprint

# Utilities
python-dotenv>=1.0
whitenoise>=6.5
django-extensions

# Testing & Quality
flake8
isort
